common: platform shim and source-level Darwin ports
common/platform.h centralizes what Darwin lacks or places elsewhere:
chimera_getrandom() (getentropy vs getrandom), sysmacros, the two
FALLOC_FL modes chimera's POSIX shim accepts, O_DIRECT/fdatasync
gaps, and printf-format helpers.  Sources swap Linux-only headers for
it and route randomness through the shim.

The POSIX layer's chimera_vfs_error is a Linux-errno-numbered protocol
enum; returning it as-is misnumbers errors on hosts whose errno values
differ (Darwin: 36 is EINPROGRESS, not ENAMETOOLONG).  Map every
protocol code to the host errno macro instead -- which also fixes
EBADF arriving as 77 (EBADFD) on Linux itself.

Per-thread impersonation (setfsuid/setgroups) is Linux-only; off Linux
chimera_setup_credential fails closed with EPERM rather than acting as
the server identity.  Only the linux and io_uring backends reach it,
and neither builds elsewhere.

REST's crypt_r() gate widens from __linux__ to CHIMERA_HAVE_XCRYPT so
the libxcrypt build verifies full crypt(5) schemes on Darwin.

+++ src/vfs/cairn/cairn.c
@@ -9,7 +9,7 @@
 #include <time.h>
 #include <unistd.h>
 #include <sys/stat.h>
-#include <endian.h>
+#include "common/platform.h"
 #include <rocksdb/c.h>
 #include "rocksdb_compat.h"
 
@@ -3312,7 +3312,7 @@ cairn_read(
         }
 
         uint64_t extent_start = be64toh(extent_key->offset);
-        uint64_t extent_length;
+        size_t   extent_length;
         rocksdb_iter_value(iter, &extent_length);
         uint64_t extent_end = extent_start + extent_length;
 
@@ -3445,7 +3445,7 @@ cairn_punch_hole(
 
     while (rocksdb_iter_valid(iter)) {
         extent_key = (struct cairn_extent_key *) rocksdb_iter_key(iter, &klen);
-        uint64_t    extent_length;
+        size_t      extent_length;
         const char *extent_data = rocksdb_iter_value(iter, &extent_length);
 
         // Stop if we've moved past this inode
